backend deletPost controller edited. Admin can also delete the post

+++ backend/controllers/post.js
@@ -129,6 +129,22 @@ exports.deletePost = async (req, res) => {
   try {
     const { userId, postId } = req.body;
 
+    //see if user is admin
+    const isAdmin = await pool.query(
+      'SELECT admin FROM users WHERE user_id = $1',
+      [userId]
+    );
+
+    //is user is the OP
+    const isOP = await pool.query(
+      'SELECT post_id, user_id FROM posts WHERE post_id = $1 AND user_id = $2',
+      [postId, userId]
+    );
+
+    if (!isAdmin.rows[0].admin && isOP.rows.length === 0) {
+      return res.status(401).json({ message: 'User has no authorisation to delete this post' });
+    }
+
     //delete all comments related to the post (table of comment used post_id as foreign keys)
     await pool.query('DELETE FROM comments WHERE post_id = $1 RETURNING *', [
       postId,
